rewrite test to just handle initial key-upload, set 0 correctly
previously the test should test if the client reacts on different
counts. For simplicity we now just check if a sync-response triggers
just a key-upload - without asking for the count beforehead.

when there is no key on the server, the server responds an empty
device_one_time_keys_count object and not 0. Now this case is handled