Remove unused ICE candidate trtrying logic
This was never used because the code which I've also added a
comment to swallows the error that it would otherwise have caught.

We should retry these candidates, but the js-sdk already has
backoff based retrying in scheduler. It only queues / retries
m.room.message events currently, but we should probably use this
for ICE candidates too.